move container fixture into testing package for reusability
