+++ localstack/testing/pytest/bootstrap.py
@@ -1,0 +1,8 @@
+import pytest
+
+from localstack import config
+
+
+@pytest.fixture(scope="session", autouse=True)
+def setup_host_config_dirs():
+    config.dirs.mkdirs()
